Move GainAbilitySpellsEffect to common place
