+++ Mage.Sets/src/mage/cards/f/FiresongAndSunspeaker.java
@@ -32,14 +32,12 @@ import java.util.UUID;
 import mage.MageInt;
 import mage.MageObject;
 import mage.ObjectColor;
-import mage.abilities.Ability;
 import mage.abilities.TriggeredAbilityImpl;
 import mage.abilities.common.SimpleStaticAbility;
-import mage.abilities.effects.ContinuousEffectImpl;
 import mage.abilities.effects.Effect;
+import mage.abilities.effects.GainAbilitySpellsEffect;
 import mage.abilities.effects.common.DamageTargetEffect;
 import mage.abilities.keyword.LifelinkAbility;
-import mage.cards.Card;
 import mage.constants.*;
 import mage.cards.CardImpl;
 import mage.cards.CardSetInfo;
@@ -49,10 +47,7 @@ import mage.filter.predicate.mageobject.CardTypePredicate;
 import mage.filter.predicate.mageobject.ColorPredicate;
 import mage.game.Game;
 import mage.game.events.GameEvent;
-import mage.game.permanent.Permanent;
 import mage.game.stack.Spell;
-import mage.game.stack.StackObject;
-import mage.players.Player;
 import mage.target.common.TargetCreatureOrPlayer;
 
 /**
@@ -131,69 +126,5 @@ class FiresongAndSunspeakerTriggeredAbility extends TriggeredAbilityImpl {
     @Override
     public String getRule() {
         return "Whenever a white instant or sorcery spell causes you to gain life, Firesong and Sunspeaker deals 3 damage to target creature or player.";
-    }
-}
-
-class GainAbilitySpellsEffect extends ContinuousEffectImpl {
-
-    private final Ability ability;
-    private final FilterObject filter;
-
-    public GainAbilitySpellsEffect(Ability ability, FilterObject filter) {
-        super(Duration.Custom, Layer.AbilityAddingRemovingEffects_6, SubLayer.NA, Outcome.AddAbility);
-        this.ability = ability;
-        this.filter = filter;
-        staticText = filter.getMessage() + " have " + ability.getRule();
-    }
-
-    public GainAbilitySpellsEffect(final GainAbilitySpellsEffect effect) {
-        super(effect);
-        this.ability = effect.ability;
-        this.filter = effect.filter;
-    }
-
-    @Override
-    public GainAbilitySpellsEffect copy() {
-        return new GainAbilitySpellsEffect(this);
-    }
-
-    @Override
-    public boolean apply(Game game, Ability source) {
-        Player player = game.getPlayer(source.getControllerId());
-        Permanent permanent = game.getPermanent(source.getSourceId());
-        if (player != null && permanent != null) {
-            for (Card card : game.getExile().getAllCards(game)) {
-                if (card.getOwnerId().equals(source.getControllerId()) && filter.match(card, game)) {
-                    game.getState().addOtherAbility(card, ability);
-                }
-            }
-            for (Card card : player.getLibrary().getCards(game)) {
-                if (filter.match(card, game)) {
-                    game.getState().addOtherAbility(card, ability);
-                }
-            }
-            for (Card card : player.getHand().getCards(game)) {
-                if (filter.match(card, game)) {
-                    game.getState().addOtherAbility(card, ability);
-                }
-            }
-            for (Card card : player.getGraveyard().getCards(game)) {
-                if (filter.match(card, game)) {
-                    game.getState().addOtherAbility(card, ability);
-                }
-            }
-            for (StackObject stackObject : game.getStack()) {
-                if (stackObject.getControllerId().equals(source.getControllerId())) {
-                    Card card = game.getCard(stackObject.getSourceId());
-                    if (card != null && filter.match(card, game)) {
-                        if (!card.getAbilities().contains(ability)) {
-                            game.getState().addOtherAbility(card, ability);
-                        }
-                    }
-                }
-            }
-            return true;
-        }
-        return false;
     }
 }
